Dashboards: Extract panel id helpers to break circular imports
