+++ public/app/features/alerting/unified/utils/rule-form.ts
@@ -17,11 +17,8 @@ import { type VizPanel, sceneGraph } from '@grafana/scenes';
 import { type DataQuery, type DataSourceRef } from '@grafana/schema';
 import { type DashboardModel } from 'app/features/dashboard/state/DashboardModel';
 import { type PanelModel } from 'app/features/dashboard/state/PanelModel';
-import {
-  getDashboardSceneFor,
-  getPanelIdForVizPanel,
-  getQueryRunnerFor,
-} from 'app/features/dashboard-scene/utils/utils';
+import { getDashboardSceneFor, getQueryRunnerFor } from 'app/features/dashboard-scene/utils/utils';
+import { getPanelIdForVizPanel } from 'app/features/dashboard-scene/utils/utils-panels';
 import { ExpressionDatasourceUID, type ExpressionQuery, ExpressionQueryType } from 'app/features/expressions/types';
 import { getTemplateSrv } from 'app/features/templating/template_srv';
 import { type RuleWithLocation } from 'app/types/unified-alerting';
