+++ public/app/features/dashboard-scene/mutation-api/commands/listPanels.ts
@@ -17,7 +17,7 @@ import { type DataFrame, type DataQueryError, LoadingState } from '@grafana/data
 import { sceneGraph, SceneDataTransformer, type SceneObject, type VizPanel } from '@grafana/scenes';
 
 import { getElements } from '../../serialization/layoutSerializers/utils';
-import { getVizPanelKeyForPanelId } from '../../utils/utils';
+import { getVizPanelKeyForPanelId } from '../../utils/utils-panels';
 import type {
   FrameSchema,
   FieldSchema,
