+++ public/app/features/dashboard-scene/panel-edit/PanelEditNext/QueryEditor/hooks/useAlertRulesForPanel.test.ts
@@ -5,7 +5,8 @@ import { type VizPanel } from '@grafana/scenes';
 import { PromAlertingRuleState, PromRuleType } from 'app/types/unified-alerting-dto';
 
 import { promAlertStateToAlertState } from '../../../../scene/AlertStatesDataLayer';
-import { getDashboardSceneFor, getPanelIdForVizPanel } from '../../../../utils/utils';
+import { getDashboardSceneFor } from '../../../../utils/utils';
+import { getPanelIdForVizPanel } from '../../../../utils/utils-panels';
 import { type PanelDataPaneNext } from '../../PanelDataPaneNext';
 
 import { useAlertRulesForPanel } from './useAlertRulesForPanel';
@@ -14,6 +15,7 @@ import { useAlertRulesForPanel } from './useAlertRulesForPanel';
 jest.mock('app/features/alerting/unified/hooks/usePanelCombinedRules');
 jest.mock('../../../../scene/AlertStatesDataLayer');
 jest.mock('../../../../utils/utils');
+jest.mock('../../../../utils/utils-panels');
 
 const mockUsePanelCombinedRules = jest.fn();
 const mockGetDashboardSceneFor = getDashboardSceneFor as jest.MockedFunction<typeof getDashboardSceneFor>;
