+++ public/app/features/dashboard-scene/panel-edit/PanelEditNext/QueryEditor/hooks/useAlertRulesForPanel.ts
@@ -6,7 +6,8 @@ import { usePanelCombinedRules } from 'app/features/alerting/unified/hooks/usePa
 import { type AlertingRule, type CombinedRule } from 'app/types/unified-alerting';
 
 import { promAlertStateToAlertState } from '../../../../scene/AlertStatesDataLayer';
-import { getDashboardSceneFor, getPanelIdForVizPanel } from '../../../../utils/utils';
+import { getDashboardSceneFor } from '../../../../utils/utils';
+import { getPanelIdForVizPanel } from '../../../../utils/utils-panels';
 import { type PanelDataPaneNext } from '../../PanelDataPaneNext';
 import { type AlertingState } from '../QueryEditorContext';
 import { type AlertRule } from '../types';
