+++ public/app/features/dashboard-scene/scene/PanelNotices.tsx
@@ -1,7 +1,7 @@
 import { type SceneComponentProps, SceneObjectBase, VizPanel, sceneGraph } from '@grafana/scenes';
 import { PanelHeaderNotices } from 'app/features/dashboard/dashgrid/PanelHeader/PanelHeaderNotices';
 
-import { getPanelIdForVizPanel } from '../utils/utils';
+import { getPanelIdForVizPanel } from '../utils/utils-panels';
 
 export class PanelNotices extends SceneObjectBase {
   static Component = PanelNoticesRenderer;
