+++ public/app/features/dashboard-scene/scene/UnconfiguredPanel.tsx
@@ -36,7 +36,8 @@ import {
   textFrames,
   useViewPhase,
 } from '../utils/unconfiguredPanelUtils';
-import { findVizPanelByKey, getVizPanelKeyForPanelId } from '../utils/utils';
+import { findVizPanelByKey } from '../utils/utils';
+import { getVizPanelKeyForPanelId } from '../utils/utils-panels';
 
 import { DashboardScene } from './DashboardScene';
