+++ public/app/features/dashboard-scene/scene/setDashboardPanelContext.ts
@@ -12,12 +12,8 @@ import { InspectTab } from 'app/features/inspector/types';
 import { openPanelInspector } from '../inspect/panelInspectorOpener';
 import { dashboardSceneGraph } from '../utils/dashboardSceneGraph';
 import { getDatasourceFromQueryRunner } from '../utils/getDatasourceFromQueryRunner';
-import {
-  getDashboardSceneFor,
-  getPanelIdForVizPanel,
-  getQueryRunnerFor,
-  isNewPanelQueryErrorsUIEnabled,
-} from '../utils/utils';
+import { getDashboardSceneFor, getQueryRunnerFor, isNewPanelQueryErrorsUIEnabled } from '../utils/utils';
+import { getPanelIdForVizPanel } from '../utils/utils-panels';
 
 import { type DashboardScene } from './DashboardScene';
