+++ public/app/features/dashboard-scene/utils/PanelModelCompatibilityWrapper.ts
@@ -2,7 +2,8 @@ import { type PanelModel } from '@grafana/data';
 import { SceneDataTransformer, type VizPanel } from '@grafana/scenes';
 import { type DataSourceRef, type DataTransformerConfig } from '@grafana/schema';
 
-import { getPanelIdForVizPanel, getQueryRunnerFor } from './utils';
+import { getQueryRunnerFor } from './utils';
+import { getPanelIdForVizPanel } from './utils-panels';
 
 export class PanelModelCompatibilityWrapper implements PanelModel {
   constructor(private _vizPanel: VizPanel) {}
