+++ public/app/features/dashboard-scene/utils/pathId.ts
@@ -1,6 +1,6 @@
 import { type SceneObject, VizPanel, sceneGraph } from '@grafana/scenes';
 
-import { getVizPanelKeyForPanelId } from './utils';
+import { getVizPanelKeyForPanelId } from './utils-panels';
 
 export function findVizPanelByPathId(scene: SceneObject, pathId: string): VizPanel | null {
   // Check if pathId is just an old legacy panel id
