+++ public/app/features/dashboard-scene/utils/utils-panels.ts
@@ -1,0 +1,13 @@
+import { type SceneObject } from '@grafana/scenes';
+
+export function getVizPanelKeyForPanelId(panelId: number) {
+  return `panel-${panelId}`;
+}
+
+export function getGridItemKeyForPanelId(panelId: number): string {
+  return `grid-item-${panelId}`;
+}
+
+export function getPanelIdForVizPanel(panel: SceneObject): number {
+  return parseInt(panel.state.key!.replace('panel-', ''), 10);
+}
