+++ public/app/features/dashboard-scene/utils/utils.ts
@@ -40,6 +40,8 @@ import { setDashboardPanelContext } from '../scene/setDashboardPanelContext';
 import { type DashboardDropTarget } from '../scene/types/DashboardDropTarget';
 import { type DashboardSceneState } from '../scene/types/dashboard';
 
+import { getVizPanelKeyForPanelId } from './utils-panels';
+
 export const NEW_PANEL_HEIGHT = 8;
 export const NEW_PANEL_WIDTH = 12;
 
@@ -48,19 +50,11 @@ const V1_PANEL_PROPERTIES = {
   COLLAPSED: 'collapsed',
 } as const;
 
-export function getVizPanelKeyForPanelId(panelId: number) {
-  return `panel-${panelId}`;
-}
-
 /**
  * Whether the new panel query errors & notices UI (header popover + dedicated inspector tab) is enabled.
  */
 export function isNewPanelQueryErrorsUIEnabled(): boolean {
   return getFeatureFlagClient().getBooleanValue(FlagKeys.GrafanaNewPanelQueryErrorsUI, false);
-}
-
-export function getPanelIdForVizPanel(panel: SceneObject): number {
-  return parseInt(panel.state.key!.replace('panel-', ''), 10);
 }
 
 /**
@@ -406,10 +400,6 @@ export function forceActivateFullSceneObjectTree(so: SceneObject): CancelActivat
  * Useful when rendering a scene object out of context of it's parent
  */
 export const activateInActiveParents = activateSceneObjectAndParentTree;
-
-export function getGridItemKeyForPanelId(panelId: number): string {
-  return `grid-item-${panelId}`;
-}
 
 export function useDashboard(scene: SceneObject): DashboardScene {
   return getDashboardSceneFor(scene);
