+++ public/app/plugins/datasource/dashboard/datasource.ts
@@ -28,11 +28,8 @@ import {
   SceneDataTransformer,
   type SceneObject,
 } from '@grafana/scenes';
-import {
-  activateSceneObjectAndParentTree,
-  findVizPanelByKey,
-  getVizPanelKeyForPanelId,
-} from 'app/features/dashboard-scene/utils/utils';
+import { activateSceneObjectAndParentTree, findVizPanelByKey } from 'app/features/dashboard-scene/utils/utils';
+import { getVizPanelKeyForPanelId } from 'app/features/dashboard-scene/utils/utils-panels';
 
 import { MIXED_REQUEST_PREFIX } from '../mixed/MixedDataSource';
